src/requirements.txt
Package                   Version
------------------------- -------------------
absl-py                   0.12.0
altgraph                  0.17
annoy                     1.16.3
argon2-cffi               20.1.0
astroid                   2.5
astunparse                1.6.3
async-generator           1.10
attrs                     21.2.0
backcall                  0.2.0
bleach                    3.3.0
cachetools                4.2.2
certifi                   2020.12.5
cffi                      1.14.5
chardet                   4.0.0
click                     8.0.0
cycler                    0.10.0
Cython                    0.29.23
decorator                 5.0.7
defusedxml                0.7.1
entrypoints               0.3
flatbuffers               1.12
gast                      0.4.0
google-auth               1.30.0
google-auth-oauthlib      0.4.4
google-pasta              0.2.0
grpcio                    1.34.1
h5py                      3.1.0
hdbscan                   0.8.26
idna                      2.10
imageio                   2.8.0
ipykernel                 5.5.4
ipython                   7.23.1
ipython-genutils          0.2.0
ipywidgets                7.6.3
isort                     4.3.21
jedi                      0.18.0
Jinja2                    3.0.0
joblib                    1.0.1
jsonschema                3.2.0
jupyter                   1.0.0
jupyter-client            6.1.12
jupyter-console           6.4.0
jupyter-core              4.7.1
jupyterlab-pygments       0.1.2
jupyterlab-widgets        1.0.0
Keras                     2.4.3
keras-nightly             2.5.0.dev2021032900
Keras-Preprocessing       1.1.2
kiwisolver                1.3.1
lazy-object-proxy         1.6.0
Markdown                  3.3.4
MarkupSafe                2.0.0
matplotlib                3.3.3
matplotlib-inline         0.1.2
mccabe                    0.6.1
mistune                   0.8.4
mtcnn                     0.1.0
nbclient                  0.5.3
nbconvert                 6.0.7
nbformat                  5.1.3
nest-asyncio              1.5.1
nltk                      3.5
notebook                  6.3.0
numpy                     1.19.5
oauthlib                  3.1.0
opencv-contrib-python     4.4.0.44
opencv-python             4.4.0.44
opt-einsum                3.3.0
packaging                 20.9
pandas                    0.25.3
pandocfilters             1.4.3
parso                     0.8.2
pexpect                   4.8.0
pickleshare               0.7.5
Pillow                    8.2.0
pip                       20.1.1
pip-tools                 5.2.1
pipdeptree                1.0.0
prometheus-client         0.10.1
prompt-toolkit            3.0.18
protobuf                  3.16.0
ptyprocess                0.7.0
pyasn1                    0.4.8
pyasn1-modules            0.2.8
pycparser                 2.20
Pygments                  2.9.0
pyinstaller               4.3
pyinstaller-hooks-contrib 2021.1
pylint                    2.5.3
pyparsing                 2.4.7
pyrsistent                0.17.3
python-dateutil           2.8.1
pytz                      2021.1
PyYAML                    5.4.1
pyzmq                     22.0.3
qtconsole                 5.1.0
QtPy                      1.9.0
rawpy                     0.14.0
regex                     2021.4.4
requests                  2.25.1
requests-oauthlib         1.3.0
rsa                       4.7.2
scikit-learn              0.24.2
scipy                     1.6.3
seaborn                   0.10.1
Send2Trash                1.5.0
setuptools                47.1.0
six                       1.15.0
sklearn                   0.0
tensorboard               2.5.0
tensorboard-data-server   0.6.1
tensorboard-plugin-wit    1.8.0
tensorflow                2.5.0
tensorflow-estimator      2.5.0
tensorflow-hub            0.8.0
tensorflow-text           2.5.0
termcolor                 1.1.0
terminado                 0.9.5
testpath                  0.4.4
threadpoolctl             2.1.0
toml                      0.10.2
tornado                   6.1
tqdm                      4.60.0
traitlets                 5.0.5
typing-extensions         3.7.4.3
urllib3                   1.26.4
wcwidth                   0.2.5
webencodings              0.5.1
Werkzeug                  2.0.0
wheel                     0.36.2
widgetsnbextension        3.5.1
wrapt                     1.12.1
